+++ CHANGELOG.md
@@ -1,1 +1,8 @@
+1.1.0
+    props: remove prop 'fnMatch'
+    methods: 
+        search(keyword) -> search(keyword, fnMatch)
+        refreshItems() -> refresh()
+        add method: sort(node, recursive, fnSort)
+
 1.0.0  the first stable version